feat: [flutter_chassis] Add multi-screen to example project.
