+++ flutter_chassis/example/pubspec.yaml
@@ -32,7 +32,6 @@ dependencies:
     sdk: flutter
   flutter_chassis:
     path: ../
-  http: ^0.13.5
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
